# Core/Accounts/CreateAccountDto.cs
using System.ComponentModel.DataAnnotations;

namespace Core.Accounts
{
    public class CreateAccountDto
    {
        [EmailAddress]
        public string Email { get; set; }

        [Required]
        public string Password { get; set; }

        [Required]
        public string ConfirmPassword { get; set; }
    }
}